init
Cursor and Gemini had fun making this for me today. Will update with actual fixes once I deploy it as I am sure it has a bit of A.I. Hallucinations  in it. Use at your own risk and all that jazz. enjoy. Welcome pull request to make it better.